chore: nice-grpc dep redirection for browser via separate package + controlling entry
