+++ ts/packages/build-scripts/getBaseConfig.ts
@@ -15,15 +15,28 @@ const BROWSERSLIST_TARGETS = browsersListToEsBuild()
   // FIXME(https://github.com/evanw/esbuild/issues/3501) Have to filter out versions like `safariTP`
   .filter((v) => !v.endsWith('TP')) as Options['target']
 
+// NOTE: about entry file overrides. We tried to solve this by setting the esbuild `options.alias`, but
+//       that did not work. Possibly tsup is not passing that along correctly.
+//       Instead we create separate pacakges (i.e. luzid-grpc-connection) which have two different entry
+//       files to then depend on browser or Node.js versions of specific dependencies.
+
+// `node` entry file which is used for Node.js and all other builds unless `ENTRY_FILE_BROWSER` is specified
 const entryFile = process.env.ENTRY_FILE
 assert(entryFile != null, 'Missing ENTRY_FILE environment variable')
-const entry = path.join('src', entryFile)
+const entryNode = path.join('src', entryFile)
+
+// Browser entry file which is used all no -`node` builds if `ENTRY_FILE_BROWSER` is specified
+const entryFileBrowser = process.env.ENTRY_FILE_BROWSER
+const entryBrowser =
+  entryFileBrowser != null ? path.join('src', entryFileBrowser) : undefined
 
 export function getBaseConfig(
   platform: Platform,
   formats: Format[],
   _options: Options
 ): Options[] {
+  const entry = platform == 'node' ? entryNode : entryBrowser ?? entryNode
+
   return [true, false]
     .flatMap<Options | null>((isDebugBuild) =>
       formats.map((format) =>
