+++ ts/packages/luzid-grpc-connection/README.md
@@ -1,0 +1,8 @@
+# @luzid/grpc-connection
+
+This is a light wrapper around the `nice-grpc` (Node.js) `nice-grpc-web` (all others)
+dependencies to provide a unified interface for gRPC connections to the @luzid/grpc-client.
+
+## LICENSE
+
+MIT
