+++ ts/packages/luzid-grpc-connection/tsconfig.declarations.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "declaration": true,
+        "declarationMap": true,
+        "emitDeclarationOnly": true,
+        "outDir": "./dist/types"
+    },
+    "extends": "./tsconfig.json",
+    "include": ["src/grpc-connection.browser.ts"]
+}
