+++ ts/packages/luzid-grpc-connection/typedoc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../typedoc.base.json"],
+  "entryPoints": ["src/grpc-client.ts"]
+}
